Fix error notyfications and refactoring
