+++ packages/xeus/src/worker.ts
@@ -18,7 +18,6 @@ import {
   removePackagesFromEmscriptenFS,
   showPackagesList
 } from '@emscripten-forge/mambajs';
-import { parseCommandLine } from './tools';
 globalThis.Module = {};
 
 // when a toplevel cell uses an await, the cell is implicitly
@@ -52,21 +51,376 @@ globalThis.ready = new Promise(resolve => {
   kernelReady = resolve;
 });
 
+namespace CommandParser {
+  export interface IParsedCommands {
+    install: IInstallationCommandOptions;
+    run: string;
+  }
+
+  export interface IInstallationCommandOptions {
+    channels?: string[];
+    specs?: string[];
+    pipSpecs?: string[];
+    isPipCommand?: boolean;
+  }
+
+  export interface ICommands extends IParsedCommands {
+    list: boolean[];
+  }
+  export type SpecTypes = 'specs' | 'pipSpecs';
+}
+
+/**
+ * Parser of installation command
+ */
+class CommandParser {
+  constructor() {}
+
+  /**
+   * Parses a command-line string and classifies it into installation commands,
+   * runnable code, or conda list operations.
+   *
+   * - If the code is a list command, it sets the `list` flag to true.
+   * - If the code contains conda or pip installation command, then it tries to parse it
+   * - Otherwise code will be executed as it is
+   *
+   * @param {string} code - The raw command-line input string to be parsed.
+   * @returns {CommandParser.ICommands} An object containing:
+   *  - parsed installation options,
+   *  - run command code,
+   *  - and a list flag indicating whether a list command was detected.
+   */
+  parseCommandLine(code: string): CommandParser.ICommands {
+    let result: CommandParser.ICommands = {
+      install: {},
+      run: code,
+      list: [false]
+    };
+    const codeLines = code.split('\n');
+    if (codeLines.length > 1) {
+      result = { ...this._parseLines(codeLines) };
+    } else {
+      if (this._hasCondaListCommand(code)) {
+        result = { install: result.install, run: '', list: [true] };
+      } else {
+        result = { ...this._parseCommand(code), list: [false] };
+      }
+    }
+    return result;
+  }
+
+  /**
+   * Parses one row of code and detects whether it is conda or pip installation command.
+   * runnable code, or conda list operations.
+   *
+   * @param {string} code - The raw command-line input string to be parsed.
+   * @returns {CommandParser.IParsedCommands} An object containing:
+   *  - parsed installation options,
+   *  - run command code
+   */
+  _parseCommand(code: string): CommandParser.IParsedCommands {
+    const run = code;
+    let isPipCommand = false;
+    const isCondaCommand = this._hasCondaInstallCommand(code);
+    code = this._isCondaCodeLine(code);
+
+    if (!isCondaCommand && code.includes('%pip install')) {
+      code = code.replace('%pip install', '');
+      isPipCommand = true;
+    }
+    let result: CommandParser.IInstallationCommandOptions = {
+      channels: [],
+      specs: [],
+      pipSpecs: []
+    };
+    if ((isCondaCommand || isPipCommand) && code) {
+      if (isPipCommand) {
+        result = this._parsePipCommand(code);
+      } else {
+        result = this._parseCondaCommand(code);
+      }
+
+      return { install: { ...result, isPipCommand }, run: '' };
+    } else {
+      return { install: {}, run };
+    }
+  }
+
+  /**
+   * Parses multiply lines
+   *
+   * @param {string[]} codeLines - The command line which should be parsed.
+   * @returns {CommandParser.ICommands} An object containing:
+   *  - parsed installation options,
+   *  - run command code,
+   *  - and a list flag indicating whether a list command was detected.
+   */
+
+  _parseLines(codeLines: string[]): CommandParser.ICommands {
+    const installCommands: string[] = [];
+    const runCommands: string[] = [];
+    const listCommand: boolean[] = [];
+
+    let channels: string[] = [];
+    let specs: string[] = [];
+    let pipSpecs: string[] = [];
+    codeLines.forEach((line: string) => {
+      if (this._hasCondaInstallCommand(line) || this._hasPipCommand(line)) {
+        installCommands.push(line);
+      } else if (this._hasCondaListCommand(line)) {
+        listCommand.push(true);
+      } else {
+        runCommands.push(line);
+      }
+    });
+
+    if (installCommands.length) {
+      let tmpResult: CommandParser.IParsedCommands = {
+        install: {
+          channels: [],
+          specs: [],
+          pipSpecs: []
+        },
+        run: ''
+      };
+      installCommands.forEach((line: string) => {
+        tmpResult = { ...this._parseCommand(line) };
+        channels = tmpResult.install.channels
+          ? [...channels, ...tmpResult.install.channels]
+          : channels;
+        specs = tmpResult.install.specs
+          ? [...specs, ...tmpResult.install.specs]
+          : specs;
+        pipSpecs = tmpResult.install.pipSpecs
+          ? [...pipSpecs, ...tmpResult.install.pipSpecs]
+          : pipSpecs;
+      });
+    }
+
+    return {
+      install: { channels, specs, pipSpecs },
+      run: runCommands ? runCommands.join('\n') : '',
+      list: listCommand ? listCommand : [false]
+    };
+  }
+
+  /**
+   * Detects whether the line has conda installation commands
+   * and replace the patter '[commandNames] install' for futher calculations
+   *
+   * @param {string} code - The command line which should be parsed.
+   * @returns {string} - Can be as part of conda installation command and as code
+   */
+  _isCondaCodeLine(code: string) {
+    const commandNames = ['micromamba', 'un', 'mamba', 'conda', 'rattler'];
+    commandNames.forEach((name: string) => {
+      if (code.includes(`%${name} install`)) {
+        code = code.replace(`%${name} install`, '');
+      }
+    });
+
+    return code;
+  }
+
+  /**
+   * Detects whether the line has conda installation commands
+   *
+   * @param {string} code - The command line which should be parsed.
+   * @returns {boolean} - True if it is a conda installation command
+   */
+  _hasCondaInstallCommand(code: string) {
+    let isCondaCommand = false;
+    const commandNames = ['micromamba', 'un', 'mamba', 'conda', 'rattler'];
+    commandNames.forEach((name: string) => {
+      if (code.includes(`%${name} install`)) {
+        isCondaCommand = true;
+      }
+    });
+
+    return isCondaCommand;
+  }
+
+  /**
+   * Detects whether the line is to list installed packages
+   *
+   * @param {string} code - The command line which should be parsed.
+   * @returns {boolean} - True if it is list command
+   */
+  _hasCondaListCommand(code: string) {
+    let isCondaListCommand = false;
+    const commandNames = ['micromamba', 'un', 'mamba', 'conda', 'rattler'];
+    commandNames.forEach((name: string) => {
+      if (code === `%${name} list`) {
+        isCondaListCommand = true;
+      }
+    });
+
+    return isCondaListCommand;
+  }
+
+  /**
+   * Detects whether the line has pip installation commands
+   *
+   * @param {string} code - The command line which should be parsed.
+   * @returns {boolean} - True if it is a pip installation command
+   */
+  _hasPipCommand(code: string) {
+    let isPipCommand = false;
+    if (code.includes('%pip install')) {
+      isPipCommand = true;
+    }
+
+    return isPipCommand;
+  }
+
+  /**
+   * Parse conda installation command
+   *
+   * @param {string} input - The command line which should be parsed.
+   * @returns {CommandParser.IInstallationCommandOptions} An object containing:
+   *  - channels,
+   *  - conda packages for installing,
+   *  - pip packages for installing
+   */
+  _parseCondaCommand = (
+    input: string
+  ): CommandParser.IInstallationCommandOptions => {
+    const parts = input.split(' ');
+    const channels: string[] = [];
+    const specs: string[] = [];
+    const pipSpecs: string[] = [];
+    for (let i = 0; i < parts.length; i++) {
+      const part = parts[i];
+      if (part) {
+        const j = i + 1;
+
+        if (part === '-c' && j < parts.length && !parts[j].startsWith('-')) {
+          channels.push(parts[j]);
+          i++;
+        } else {
+          specs.push(part);
+        }
+      }
+    }
+
+    return {
+      channels,
+      specs,
+      pipSpecs
+    };
+  };
+
+  /**
+   * Parse pip installation command
+   *
+   * @param {string} input - The command line which should be parsed.
+   * @returns {CommandParser.IInstallationCommandOptions} An object containing:
+   *  - channels,
+   *  - conda packages for installing,
+   *  - pip packages for installing
+   */
+
+  _parsePipCommand = (
+    input: string
+  ): CommandParser.IInstallationCommandOptions => {
+    const parts = input.split(' ');
+    let skip = false;
+    const limits = [
+      '--index-url',
+      '.whl',
+      'tar.gz',
+      '--extra-index-url',
+      'http',
+      'https',
+      'git',
+      './',
+      '-r',
+      '--extra-index-url'
+    ];
+
+    const flags = [
+      '--upgrade',
+      '--pre',
+      '--no-cache-dir',
+      '--user',
+      '--upgrade',
+      '--no-deps'
+    ];
+
+    const pipSpecs: string[] = [];
+
+    limits.map((options: string) => {
+      if (input.includes(options)) {
+        skip = true;
+      }
+    });
+    if (!skip) {
+      for (let i = 0; i < parts.length; i++) {
+        const part = parts[i];
+        if (part) {
+          if (!flags.includes(part)) {
+            pipSpecs.push(part);
+          }
+        }
+      }
+    }
+
+    return {
+      channels: [],
+      specs: [],
+      pipSpecs
+    };
+  };
+}
+
 export class XeusWorkerLogger implements ILogger {
   constructor(kernelId: string) {
     this._channel = new BroadcastChannel(`/kernel-broadcast/${kernelId}`);
   }
 
   log(...msg: any[]): void {
+    postMessage({
+      _stream: {
+        name: names['log'],
+        text: msg.join(' ') + '\n'
+      }
+    });
     this._channel.postMessage({ type: 'log', msg: msg.join(' ') });
   }
 
   warn(...msg: any[]): void {
+    postMessage({
+      _stream: {
+        name: names['warn'],
+        text: msg.join(' ') + '\n'
+      }
+    });
     this._channel.postMessage({ type: 'warn', msg: msg.join(' ') });
   }
 
   error(...msg: any[]): void {
-    this._channel.postMessage({ type: 'error', msg: msg.join(' ') });
+    const [firstEl, ...rest] = msg;
+    let executionCount = 0;
+    let evalue = '';
+    if (msg !== undefined) {
+      if (typeof firstEl === 'number') {
+        executionCount = firstEl;
+        evalue = rest.join('');
+      } else {
+        evalue = msg.join('');
+      }
+      postMessage({
+        _stream: {
+          name: names['error'],
+          evalue,
+          traceback: [],
+          executionCount,
+          text: evalue
+        }
+      });
+
+      this._channel.postMessage({ type: 'error', msg: msg.join(' ') });
+    }
   }
 
   private _channel: BroadcastChannel;
@@ -126,7 +480,7 @@ export abstract class XeusRemoteKernel {
   _getInstalledPackages() {
     return this._installedPackages;
   }
-
+  /*
   _getKernelLogger(): ILogger {
     return {
       log(...msg: any[]): void {
@@ -157,23 +511,20 @@ export abstract class XeusRemoteKernel {
         executionCount?: 0;
       }): void {
         if (evalue !== undefined) {
-          console.log('evalue', evalue);
-          console.log('traceback', traceback);
-          console.log('executionCount', executionCount);
           postMessage({
             _stream: {
               name: names['error'],
               evalue,
               traceback,
               executionCount,
-              text: evalue
+              text: `${evalue}${traceback?.length? traceback.join(''):''}`
             }
           });
         }
       }
     };
   }
-
+*/
   _setInstalledPackages() {
     let installed = {};
     this._empackEnvMeta.packages.map((pkg: any) => {
@@ -207,11 +558,11 @@ export abstract class XeusRemoteKernel {
       }
     });
     if (isInstallCommand || isListCommand) {
-      const kernelLogger = this._getKernelLogger();
-      const { install, run, list } = parseCommandLine(code);
+      const commandParser = new CommandParser();
+      const { install, run, list } = commandParser.parseCommandLine(code);
       const installedPackages = this._getInstalledPackages();
       if (list.includes(true)) {
-        showPackagesList(installedPackages, kernelLogger);
+        showPackagesList(installedPackages, this._logger);
       }
 
       if (install.specs || install.pipSpecs) {
@@ -220,18 +571,18 @@ export abstract class XeusRemoteKernel {
           install.pipSpecs
         );
         try {
-          kernelLogger.log(`Collecting ${packageNames?.join(',')}`);
+          this._logger.log(`Collecting ${packageNames?.join(',')}`);
           const start = performance.now();
           const newPackages = await solve({
             ymlOrSpecs: install.specs ? install.specs : [],
             installedPackages,
             pipSpecs: install.pipSpecs ? install.pipSpecs : [],
             channels: install.channels,
-            logger: kernelLogger
+            logger: this._logger
           });
           const end = performance.now();
           const time = end - start;
-          kernelLogger.log(`Solving took ${time / 1000} seconds`);
+          this._logger.log(`Solving took ${time / 1000} seconds`);
 
           await this._reloadPackages(
             {
@@ -239,17 +590,10 @@ export abstract class XeusRemoteKernel {
               ...newPackages.pipPackages
             },
             packageNames,
-            kernelLogger
+            this._logger
           );
         } catch (error: any) {
-          console.log('++++');
-          console.log('error', error);
-          kernelLogger.error({
-            evalue: error.message,
-            traceback: [error.stack],
-            execution_count: this._executionCount
-          });
-          this._logger?.error(error);
+          this._logger?.error(this._executionCount, error.stack);
         }
       }
       code = run || '';
